% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retriever-lib.R
\name{ConsolidatedDeathsData.class}
\alias{ConsolidatedDeathsData.class}
\title{ConsolidatedDeathsData.class}
\description{
ConsolidatedDeathsData.class

ConsolidatedDeathsData.class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConsolidatedDeathsData.class$new()}}
\item \href{#method-consolidate}{\code{ConsolidatedDeathsData.class$consolidate()}}
\item \href{#method-readDeathsStats}{\code{ConsolidatedDeathsData.class$readDeathsStats()}}
\item \href{#method-clone}{\code{ConsolidatedDeathsData.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsolidatedDeathsData.class$new(data.dir = getEnv("data_dir"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsolidatedDeathsData.class$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readDeathsStats"></a>}}
\if{latex}{\out{\hypertarget{method-readDeathsStats}{}}}
\subsection{Method \code{readDeathsStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsolidatedDeathsData.class$readDeathsStats(filename, fail.on.error = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsolidatedDeathsData.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
