% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19AR-lib.R
\name{COVID19ARCurator}
\alias{COVID19ARCurator}
\title{COVID19ARCurator}
\description{
COVID19ARCurator

COVID19ARCurator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARCurator$new()}}
\item \href{#method-loadData}{\code{COVID19ARCurator$loadData()}}
\item \href{#method-curateData}{\code{COVID19ARCurator$curateData()}}
\item \href{#method-makeSummary}{\code{COVID19ARCurator$makeSummary()}}
\item \href{#method-clone}{\code{COVID19ARCurator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator$new(url, data.dir = getEnv("data_dir"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-curateData"></a>}}
\if{latex}{\out{\hypertarget{method-curateData}{}}}
\subsection{Method \code{curateData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator$curateData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeSummary"></a>}}
\if{latex}{\out{\hypertarget{method-makeSummary}{}}}
\subsection{Method \code{makeSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator$makeSummary(
  group.vars = c("provincia_residencia", "edad.rango", "sepi_apertura",
    "origen_financiamiento")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
