% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19AR-lib.R
\name{COVID19ARCurator200519}
\alias{COVID19ARCurator200519}
\title{COVID19ARCurator200519}
\description{
COVID19ARCurator200519

COVID19ARCurator200519
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARCurator200519$new()}}
\item \href{#method-setupColsSpecifications}{\code{COVID19ARCurator200519$setupColsSpecifications()}}
\item \href{#method-loadData}{\code{COVID19ARCurator200519$loadData()}}
\item \href{#method-curateData}{\code{COVID19ARCurator200519$curateData()}}
\item \href{#method-curateData200519}{\code{COVID19ARCurator200519$curateData200519()}}
\item \href{#method-makeSummary}{\code{COVID19ARCurator200519$makeSummary()}}
\item \href{#method-clone}{\code{COVID19ARCurator200519$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator200519$new(url, data.dir = getEnv("data_dir"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupColsSpecifications"></a>}}
\if{latex}{\out{\hypertarget{method-setupColsSpecifications}{}}}
\subsection{Method \code{setupColsSpecifications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator200519$setupColsSpecifications()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator200519$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-curateData"></a>}}
\if{latex}{\out{\hypertarget{method-curateData}{}}}
\subsection{Method \code{curateData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator200519$curateData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-curateData200519"></a>}}
\if{latex}{\out{\hypertarget{method-curateData200519}{}}}
\subsection{Method \code{curateData200519()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator200519$curateData200519()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeSummary"></a>}}
\if{latex}{\out{\hypertarget{method-makeSummary}{}}}
\subsection{Method \code{makeSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator200519$makeSummary(
  group.vars = c("provincia_residencia", "edad.rango", "sepi_apertura",
    "origen_financiamiento")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARCurator200519$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
