% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr-lib.R
\name{COVID19ARDailyReportsPlain}
\alias{COVID19ARDailyReportsPlain}
\title{COVID19ARDailyReportsPlain}
\description{
COVID19ARDailyReportsPlain

COVID19ARDailyReportsPlain
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARDailyReportsPlain$new()}}
\item \href{#method-buildCasosMapping}{\code{COVID19ARDailyReportsPlain$buildCasosMapping()}}
\item \href{#method-buildCasosPlan}{\code{COVID19ARDailyReportsPlain$buildCasosPlan()}}
\item \href{#method-getReportUrl}{\code{COVID19ARDailyReportsPlain$getReportUrl()}}
\item \href{#method-buildMapacheData}{\code{COVID19ARDailyReportsPlain$buildMapacheData()}}
\item \href{#method-resetMerge}{\code{COVID19ARDailyReportsPlain$resetMerge()}}
\item \href{#method-buildCasosReport}{\code{COVID19ARDailyReportsPlain$buildCasosReport()}}
\item \href{#method-generateReportDaySummary}{\code{COVID19ARDailyReportsPlain$generateReportDaySummary()}}
\item \href{#method-loadReportDiffSummary}{\code{COVID19ARDailyReportsPlain$loadReportDiffSummary()}}
\item \href{#method-saveReportDiffSummary}{\code{COVID19ARDailyReportsPlain$saveReportDiffSummary()}}
\item \href{#method-clone}{\code{COVID19ARDailyReportsPlain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$new(
  min.rebuilt.date = "2020-06-16",
  report.diff.dir = "../COVID19ARdata/sources/COVID19AR"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCasosMapping"></a>}}
\if{latex}{\out{\hypertarget{method-buildCasosMapping}{}}}
\subsection{Method \code{buildCasosMapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$buildCasosMapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCasosPlan"></a>}}
\if{latex}{\out{\hypertarget{method-buildCasosPlan}{}}}
\subsection{Method \code{buildCasosPlan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$buildCasosPlan(casos.dir = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReportUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getReportUrl}{}}}
\subsection{Method \code{getReportUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$getReportUrl(commit.id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildMapacheData"></a>}}
\if{latex}{\out{\hypertarget{method-buildMapacheData}{}}}
\subsection{Method \code{buildMapacheData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$buildMapacheData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetMerge"></a>}}
\if{latex}{\out{\hypertarget{method-resetMerge}{}}}
\subsection{Method \code{resetMerge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$resetMerge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCasosReport"></a>}}
\if{latex}{\out{\hypertarget{method-buildCasosReport}{}}}
\subsection{Method \code{buildCasosReport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$buildCasosReport(max.n = 0, min.date = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateReportDaySummary"></a>}}
\if{latex}{\out{\hypertarget{method-generateReportDaySummary}{}}}
\subsection{Method \code{generateReportDaySummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$generateReportDaySummary(update.date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadReportDiffSummary"></a>}}
\if{latex}{\out{\hypertarget{method-loadReportDiffSummary}{}}}
\subsection{Method \code{loadReportDiffSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$loadReportDiffSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveReportDiffSummary"></a>}}
\if{latex}{\out{\hypertarget{method-saveReportDiffSummary}{}}}
\subsection{Method \code{saveReportDiffSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$saveReportDiffSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDailyReportsPlain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
