% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-lib.R
\name{COVID19ARLegacyCompressor}
\alias{COVID19ARLegacyCompressor}
\title{COVID19ARLegacyCompressor}
\description{
COVID19ARLegacyCompressor

COVID19ARLegacyCompressor
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARLegacyCompressor$new()}}
\item \href{#method-buildCasosMapping}{\code{COVID19ARLegacyCompressor$buildCasosMapping()}}
\item \href{#method-getReportUrl}{\code{COVID19ARLegacyCompressor$getReportUrl()}}
\item \href{#method-compressSnapshot}{\code{COVID19ARLegacyCompressor$compressSnapshot()}}
\item \href{#method-clone}{\code{COVID19ARLegacyCompressor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARLegacyCompressor$new(
  report.source.dir = "~/.R/COVID19AR/",
  report.snapshots.dir = "../COVID19ARdata/sources/msalnacion"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCasosMapping"></a>}}
\if{latex}{\out{\hypertarget{method-buildCasosMapping}{}}}
\subsection{Method \code{buildCasosMapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARLegacyCompressor$buildCasosMapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReportUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getReportUrl}{}}}
\subsection{Method \code{getReportUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARLegacyCompressor$getReportUrl(commit.id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compressSnapshot"></a>}}
\if{latex}{\out{\hypertarget{method-compressSnapshot}{}}}
\subsection{Method \code{compressSnapshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARLegacyCompressor$compressSnapshot(snapshot.date, dest.filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARLegacyCompressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
