% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-lib.R
\name{COVID19ARSampleGenerator}
\alias{COVID19ARSampleGenerator}
\title{COVID19ARSampleGenerator}
\description{
COVID19ARSampleGenerator

COVID19ARSampleGenerator
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARSampleGenerator$new()}}
\item \href{#method-genSample}{\code{COVID19ARSampleGenerator$genSample()}}
\item \href{#method-getCaseFileName}{\code{COVID19ARSampleGenerator$getCaseFileName()}}
\item \href{#method-genSampleBatch}{\code{COVID19ARSampleGenerator$genSampleBatch()}}
\item \href{#method-clone}{\code{COVID19ARSampleGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARSampleGenerator$new(
  daily.reports,
  sample.name,
  sample.ratio = 0.01,
  seed = 0,
  output.dir = "../COVID19ARdata/dev/samples",
  min.date = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-genSample"></a>}}
\if{latex}{\out{\hypertarget{method-genSample}{}}}
\subsection{Method \code{genSample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARSampleGenerator$genSample(current.case)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCaseFileName"></a>}}
\if{latex}{\out{\hypertarget{method-getCaseFileName}{}}}
\subsection{Method \code{getCaseFileName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARSampleGenerator$getCaseFileName(current.case)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-genSampleBatch"></a>}}
\if{latex}{\out{\hypertarget{method-genSampleBatch}{}}}
\subsection{Method \code{genSampleBatch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARSampleGenerator$genSampleBatch(max.n = 0, min.date = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARSampleGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
