% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19AR-lib.R
\name{EdadCoder}
\alias{EdadCoder}
\title{EdadCoder}
\description{
EdadCoder

EdadCoder
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EdadCoder$new()}}
\item \href{#method-setupCoder}{\code{EdadCoder$setupCoder()}}
\item \href{#method-codeEdad}{\code{EdadCoder$codeEdad()}}
\item \href{#method-clone}{\code{EdadCoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupCoder"></a>}}
\if{latex}{\out{\hypertarget{method-setupCoder}{}}}
\subsection{Method \code{setupCoder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$setupCoder()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codeEdad"></a>}}
\if{latex}{\out{\hypertarget{method-codeEdad}{}}}
\subsection{Method \code{codeEdad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$codeEdad(edad)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
