% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19AR-lib.R
\name{exportAggregatedTables}
\alias{exportAggregatedTables}
\title{exportAggregatedTables}
\usage{
exportAggregatedTables(
  covid.ar.curator,
  output.dir,
  aggrupation.criteria = list(provincia_residencia = c("provincia_residencia"),
    provincia_residencia_sexo = c("provincia_residencia", "sexo"), edad_rango_sexo =
    c("edad.rango", "sexo"), provincia_residencia_sepi_apertura =
    c("provincia_residencia", "sepi_apertura"),
    provincia_departamento_residencia_sepi_apertura = c("residencia_provincia_nombre",
    "residencia_departamento_nombre", "sepi_apertura"),
    provincia_residencia_fecha_apertura = c("provincia_residencia", "fecha_apertura")),
  data2process = covid.ar.curator$getData(),
  file.prefix = "covid19ar_"
)
}
\description{
exportAggregatedTables
}
\author{
kenarab
}
