% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{fixEncoding}
\alias{fixEncoding}
\title{fixEncoding return filepath with encoding in UTF8}
\usage{
fixEncoding(file.path)
}
\description{
fixEncoding return filepath with encoding in UTF8
}
\author{
kenarab
}
