% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lib.R
\name{getDepartamentosExponentialGrowthPlot}
\alias{getDepartamentosExponentialGrowthPlot}
\title{getDepartamentosExponentialGrowthPlot}
\usage{
getDepartamentosExponentialGrowthPlot(
  covid19ar.curator,
  ma.n = 7,
  n.highlighted = 10
)
}
\description{
getDepartamentosExponentialGrowthPlot
}
