% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lib.R
\name{ReportGeneratorCOVID19AR}
\alias{ReportGeneratorCOVID19AR}
\title{ReportGeneratorCOVID19AR}
\description{
ReportGeneratorCOVID19AR

ReportGeneratorCOVID19AR
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportGeneratorCOVID19AR$new()}}
\item \href{#method-preprocess}{\code{ReportGeneratorCOVID19AR$preprocess()}}
\item \href{#method-getDepartamentosExponentialGrowthPlot}{\code{ReportGeneratorCOVID19AR$getDepartamentosExponentialGrowthPlot()}}
\item \href{#method-getDepartamentosCrossSectionConfirmedPostitivyPlot}{\code{ReportGeneratorCOVID19AR$getDepartamentosCrossSectionConfirmedPostitivyPlot()}}
\item \href{#method-clone}{\code{ReportGeneratorCOVID19AR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorCOVID19AR$new(covid19ar.curator, ma.n = 7)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorCOVID19AR$preprocess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDepartamentosExponentialGrowthPlot"></a>}}
\if{latex}{\out{\hypertarget{method-getDepartamentosExponentialGrowthPlot}{}}}
\subsection{Method \code{getDepartamentosExponentialGrowthPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorCOVID19AR$getDepartamentosExponentialGrowthPlot(
  n.highlighted = 10
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDepartamentosCrossSectionConfirmedPostitivyPlot"></a>}}
\if{latex}{\out{\hypertarget{method-getDepartamentosCrossSectionConfirmedPostitivyPlot}{}}}
\subsection{Method \code{getDepartamentosCrossSectionConfirmedPostitivyPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorCOVID19AR$getDepartamentosCrossSectionConfirmedPostitivyPlot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorCOVID19AR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
