% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths-lib.R
\name{retrieveArgentinasDeathsStatistics}
\alias{retrieveArgentinasDeathsStatistics}
\title{retrieveArgentinasDeathsStatistics}
\usage{
retrieveArgentinasDeathsStatistics()
}
\description{
retrieveArgentinasDeathsStatistics
}
