% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{retrieveURL}
\alias{retrieveURL}
\title{retrieveURL}
\usage{
retrieveURL(
  data.url,
  col.types,
  dest.dir = getEnv("data_dir"),
  force.download = FALSE,
  download.new.data = TRUE,
  daily.update.time = "20:00:00"
)
}
\description{
retrieveURL
}
\author{
kenarab
}
