% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-lib.R
\name{TestCaseCOVID19ARGenerator}
\alias{TestCaseCOVID19ARGenerator}
\title{TestCaseCOVID19ARGenerator}
\description{
TestCaseCOVID19ARGenerator

TestCaseCOVID19ARGenerator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TestCaseCOVID19ARGenerator$new()}}
\item \href{#method-generateTestCase}{\code{TestCaseCOVID19ARGenerator$generateTestCase()}}
\item \href{#method-loadCurator}{\code{TestCaseCOVID19ARGenerator$loadCurator()}}
\item \href{#method-setCurator}{\code{TestCaseCOVID19ARGenerator$setCurator()}}
\item \href{#method-applyFilters}{\code{TestCaseCOVID19ARGenerator$applyFilters()}}
\item \href{#method-getCaseFilename}{\code{TestCaseCOVID19ARGenerator$getCaseFilename()}}
\item \href{#method-generateExpectedSummaries}{\code{TestCaseCOVID19ARGenerator$generateExpectedSummaries()}}
\item \href{#method-clone}{\code{TestCaseCOVID19ARGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$new(
  name,
  testcase.dir,
  agg.criteria,
  filters = NULL,
  url.dir = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateTestCase"></a>}}
\if{latex}{\out{\hypertarget{method-generateTestCase}{}}}
\subsection{Method \code{generateTestCase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$generateTestCase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadCurator"></a>}}
\if{latex}{\out{\hypertarget{method-loadCurator}{}}}
\subsection{Method \code{loadCurator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$loadCurator()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCurator"></a>}}
\if{latex}{\out{\hypertarget{method-setCurator}{}}}
\subsection{Method \code{setCurator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$setCurator(curator)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-applyFilters"></a>}}
\if{latex}{\out{\hypertarget{method-applyFilters}{}}}
\subsection{Method \code{applyFilters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$applyFilters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCaseFilename"></a>}}
\if{latex}{\out{\hypertarget{method-getCaseFilename}{}}}
\subsection{Method \code{getCaseFilename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$getCaseFilename()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateExpectedSummaries"></a>}}
\if{latex}{\out{\hypertarget{method-generateExpectedSummaries}{}}}
\subsection{Method \code{generateExpectedSummaries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$generateExpectedSummaries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseCOVID19ARGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
