% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19AR-lib.R
\name{COVID19ARDiff}
\alias{COVID19ARDiff}
\title{COVID19ARDiff}
\description{
COVID19ARDiff

COVID19ARDiff
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARDiff$new()}}
\item \href{#method-loadReports}{\code{COVID19ARDiff$loadReports()}}
\item \href{#method-processDiff}{\code{COVID19ARDiff$processDiff()}}
\item \href{#method-saveReportDiff}{\code{COVID19ARDiff$saveReportDiff()}}
\item \href{#method-clone}{\code{COVID19ARDiff$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiff$new(min.rebuilt.date, report.diff.dir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadReports"></a>}}
\if{latex}{\out{\hypertarget{method-loadReports}{}}}
\subsection{Method \code{loadReports()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiff$loadReports(current.case, curate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-processDiff"></a>}}
\if{latex}{\out{\hypertarget{method-processDiff}{}}}
\subsection{Method \code{processDiff()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiff$processDiff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveReportDiff"></a>}}
\if{latex}{\out{\hypertarget{method-saveReportDiff}{}}}
\subsection{Method \code{saveReportDiff()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiff$saveReportDiff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
