% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19AR-lib.R
\name{COVID19ARDiffSummarizer}
\alias{COVID19ARDiffSummarizer}
\title{COVID19ARDiffBuilder}
\description{
COVID19ARDiffBuilder

COVID19ARDiffBuilder
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARDiffSummarizer$new()}}
\item \href{#method-buildCasosMapping}{\code{COVID19ARDiffSummarizer$buildCasosMapping()}}
\item \href{#method-buildMapacheData}{\code{COVID19ARDiffSummarizer$buildMapacheData()}}
\item \href{#method-buildCasosReport}{\code{COVID19ARDiffSummarizer$buildCasosReport()}}
\item \href{#method-generateReportDaySummary}{\code{COVID19ARDiffSummarizer$generateReportDaySummary()}}
\item \href{#method-loadReportDiffSummary}{\code{COVID19ARDiffSummarizer$loadReportDiffSummary()}}
\item \href{#method-saveReportDiffSummary}{\code{COVID19ARDiffSummarizer$saveReportDiffSummary()}}
\item \href{#method-clone}{\code{COVID19ARDiffSummarizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$new(
  min.rebuilt.date = "2020-06-16",
  report.diff.dir = "../COVID19ARdata/sources/COVID19AR"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCasosMapping"></a>}}
\if{latex}{\out{\hypertarget{method-buildCasosMapping}{}}}
\subsection{Method \code{buildCasosMapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$buildCasosMapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildMapacheData"></a>}}
\if{latex}{\out{\hypertarget{method-buildMapacheData}{}}}
\subsection{Method \code{buildMapacheData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$buildMapacheData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCasosReport"></a>}}
\if{latex}{\out{\hypertarget{method-buildCasosReport}{}}}
\subsection{Method \code{buildCasosReport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$buildCasosReport(max.n = 0, min.date = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateReportDaySummary"></a>}}
\if{latex}{\out{\hypertarget{method-generateReportDaySummary}{}}}
\subsection{Method \code{generateReportDaySummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$generateReportDaySummary(update.date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadReportDiffSummary"></a>}}
\if{latex}{\out{\hypertarget{method-loadReportDiffSummary}{}}}
\subsection{Method \code{loadReportDiffSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$loadReportDiffSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveReportDiffSummary"></a>}}
\if{latex}{\out{\hypertarget{method-saveReportDiffSummary}{}}}
\subsection{Method \code{saveReportDiffSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$saveReportDiffSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARDiffSummarizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
