% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{zipFile}
\alias{zipFile}
\title{zipFile}
\usage{
zipFile(
  source.dir,
  current.file,
  dest.dir = source.dir,
  flags = "",
  rm.original = TRUE,
  overwrite = FALSE
)
}
\description{
zipFile
}
\author{
kenarab
}
