% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{zipFile}
\alias{zipFile}
\title{zipFile}
\usage{
zipFile(
  home.dir,
  current.file,
  rm.original = TRUE,
  overwrite = FALSE,
  minimum.size.accepted = 2000
)
}
\description{
zipFile
}
\author{
kenarab
}
